requirements.txt
streamlit==0.64.0
tensorflow==2.0
opencv-contrib-python-headless
Pillow==7.1.2
numpy==1.19.0
scipy==1.5.2